+++ hydra-node/test/Hydra/OffChainAgreementSpec.hs
@@ -105,6 +105,7 @@ import Hydra.Tx.IsTx (TxIdType, UTxOType)
 import Hydra.Tx.Party (Party)
 import Hydra.Tx.Snapshot (ConfirmedSnapshot (..), Snapshot (..))
 import Test.Hydra.Ledger.Simple (utxoRef)
+import Test.Hydra.Node.Fixture (testRollbackHorizon)
 import Test.Hydra.Tx.Fixture (alice, aliceSk, bob, bobSk, carol, carolSk, deriveOnChainId, testHeadId)
 import Test.QuickCheck (choose, conjoin, counterexample, elements, forAll, sublistOf, (===))
 
@@ -122,6 +123,7 @@ aliceEnv =
     , depositPeriod = defaultDepositPeriod
     , depositActivation = defaultDepositActivation
     , unsyncedPeriod = defaultUnsyncedPeriod
+    , rollbackHorizon = testRollbackHorizon
     , participants = deriveOnChainId <$> threeParties
     , configuredPeers = ""
     }